aieng-eval-agents/pyproject.toml
[project]
name = "aieng-eval-agents"
version = "0.2.0"
description = "Agentic AI Evaluation reference implementation from Vector Institute AI Engineering"
authors = [{name = "Vector AI Engineering", email = "ai_engineering@vectorinstitute.ai"}]
requires-python = ">=3.12,<4.0"
readme = "README.md"
license = "MIT"
dependencies = [
    "kagglehub>=0.4.1",
    "pydantic-settings>=2.12.0",
    "pydantic>=2.12.0",
    "google-genai>=1.52.0",
    "google-adk>=1.23.0",
    "pandas>=2.3.0",
    "rich>=13.9.0",
    "langfuse>=3.10.3",
    "openinference-instrumentation-google-adk>=0.1.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "pypdf",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
knowledge-agent = "aieng.agent_evals.knowledge_agent.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["aieng/"]

[tool.hatch.build.targets.wheel]
include = ["aieng/"]
